Fixes hiding of disclaimer view
Since we added a spacer via the stackview.addSpacer extension helper method instead of the NSOnboardingContentViewController (which wraps any view in a WrapperView before adding it to the stackView), when performing an onboarding step the NSOnboardingContentViewController.fadeAnimation would not call the completion handler since the last stackview element (= the spacer view) does not have any subviews.

As a consequence, the DisclaimerViewController View is not set to hidden after an onboarding step is taken. This behaviour is only visible either in the view hierarchy inspector, or when running voiceover (since this would read out some of the hidden elements, one of them being the privacy info button for example) .

As a solution, we need to add a spacer view via the addArrangedView of the NSOnboardingContentViewController.