Use new .authorizationUnknown state to distinguish whether user denied permission or was not yet prompted
If user has not been prompted yet and taps on "Activate" we will
startTracing() which in turn calls setExposureNotificationEnabled() which
then results in the authorization prompt.

This is necessary after a user presses "Turn Off Exposure Notifications"
in the iOS Settings.app, which revokes permissions and resets
authorization to .unknown.